get_size returns u64

+++ condow_core/src/condow_client.rs
@@ -35,7 +35,7 @@ pub trait CondowClient: Clone + Send + Sync + 'static {
     type Location: std::fmt::Debug + std::fmt::Display + Clone + Send + Sync + 'static;
 
     /// Returns the size of the BLOB at the given location
-    fn get_size(&self, location: Self::Location) -> BoxFuture<'static, Result<usize, CondowError>>;
+    fn get_size(&self, location: Self::Location) -> BoxFuture<'static, Result<u64, CondowError>>;
 
     /// Download a BLOB or part of a BLOB from the given location as specified by the [DownloadSpec]
     ///
